Expand the capabilities of ai
Now, instead of only generating a bash command to do something,
it utilizes the openai function calling feature to either put a bash
command on the command buffer, or to just echo something to the
terminal.

Hopefully this forms as a basis to also include other functions, like
image generation.