Migrate from pure shell to go interop
I've been developing this app further in a separate repo (Aaronik/ai).
It uses go as the main driver. This has helped primarily in two ways:

1) It makes it much easier to test. Having a fully fledged programming
   language has allowed me to write much more detailed tests.
2) It handles weird JSON much better than jq. Over the course of this
   whole project, coercing jq to accept openai's json has been a pain.
   Before this commit, many responses errored out and gave a bunch of
   gross stuff in the terminal. Now it all works swimmingly.

It does require `go` to be installed on a person's computer now, which
is certainly a limitation. But overall it's much more stable and robust,
so I think it's worthwhile.